requirements.txt
appdirs==1.4.3
future==0.16.0
httplib2==0.10.3
oauthlib==2.0.1
packaging==16.8
pyparsing==2.2.0
python-oauth2==1.0.1
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
